Add `strymon terminate` sub-command
This extends executors with the capability to send a termination
signal to their children on request. Termination requests are
sent to the coordinator and forwared to all executors participating
in a job.

On Unix, this results in a SIGTERM signal being sent to the child
process. On Windows, this functionality is currently not implemented,
but likely be done via `TerminateProcess`.

There is currently no restriction on who is allowed to terminate
a running job.

Closes #12

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>