Introduce strymon_job crate
This is a new crate intended to become the new Strymon "client-library".
It is intended to eventually replace the `strymon_runtime::query` module
and contains the newly implemented publish-subscribe protocol which
maintains progress tracking information for both the upper and lower
frontier.

For more details about the rationale behind the new protocol, please refer
to the following post on the mailing-list:

 https://lists.inf.ethz.ch/mailman/private/strymon-users/2017-December/000005.html

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>