strymon_job: Fix filter bug in subscriber logic.
This fixes a bug where elements were removed from the filter too early.
The old implementation removed elements which became part of the frontier.
This was wrong, as elements in the frontier are still part of the `active`
set and thus can still be observed. Only elements strictly smaller than
the frontier can be removed from the filter.

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>